fed_server/script/lab_requirements.txt
pillow==9.5.0
h5py==3.8.0
tensorflow-hub==0.13.0